requirements.txt
streamlit
plexapi
pandas
datetime
python-dotenv